greybus: stash hd as context for all URBs
This changes the context value stashed with each USB URB so that it
is always the host device pointer.

In cport_out_callback() this allows us to get away with *not*
requiring the gbuf for handling completions any more.  We are
(currently) ignoring the gbuf status value returned anyway, so
we'll skip setting it altogether.

Greg's comments in cport_out_callback() point out that ignoring
this was misguided, and handling send errors will be put in
place in an upcoming patch.

The context is set to the host device pointer for SVC receive and
CPort receive URBs for consistency--because we can.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>