greybus: only record message payload size
An asynchronous operation will want to know how big the response
message it receives is.  Rather than require the sender to record
that information, expose a new field "payload_size" available to
the protocol code for this purpose.

An operation message consists of a header and a payload.  The size
of the message can be derived from the size of the payload, so
record only the payload size and not the size of the whole message.
Reorder the fields in a message structure.

Update the description of the message header structure.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>